backend/requirements.txt
Flask
pymongo
Flask-Bcrypt
PyJWT
python-dotenv
uvicorn
asgiref
flasgger
flask-cors
